requirements.txt
# Core Dependencies
fastapi==0.95.2
uvicorn[standard]
sqlalchemy==2.0.16
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart

# Image Processing & OCR
Pillow==9.5.0
pytesseract==0.3.10
# PaddleOCR==2.7.0.3  # Optional alternative to Tesseract

# Testing & Quality
pytest==7.3.1
pytest-cov==4.1.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1

# Linting & Formatting
flake8==6.0.0
pylint==2.17.4
black==23.3.0
isort==5.12.0

# Utilities
requests==2.31.0
croniter==1.3.14

# Development
ipython==8.14.0
